requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
h11==0.14.0
idna==3.6
numpy==1.26.3
outcome==1.3.0.post0
pillow==10.2.0
pycparser==2.21
PySocks==1.7.1
requests==2.31.0
selenium==4.17.2
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soundfile==0.12.1
soupsieve==2.5
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
undetected-chromedriver==3.5.4
urllib3==2.1.0
websockets==12.0
wsproto==1.2.0
